wip react

+++ api/index.js
@@ -15,24 +15,32 @@ const apiport = 8000;
 const port = 8001;
 
 const mime = require('mime-types')
+const cors = require('cors');
 
-// const cors = require('cors');
+const API = require('./apiAuth');
 
 app.use(express.json());
 io.origins("*:*");
-// app.use(cors({
-//   origin: '*'
-// }));
+app.use(cors({
+  origin: '*'
+}));
+
+app.post('/api/register', (req, res) => {
+  //create a new with "user:Username"
+  let username = req.body.username;
+  let user = API.createUser(username, req);
+  res.status(201).send({ data: user });
+});
 
 app.use(express.json()); //Notice express.json middleware
 app.use(fileUpload());
 
 const pool = mariadb.createPool({
-  // host: 'localhost',
-  // port: 3306
+  host: 'localhost',
+  port: 3306,
   user: "root",
-  password: "root",
-  database: "clipboard",
+  password: "q47#YA",
+  database: "private_clipboard",
   connectionLimit: 100,
 });
 
@@ -52,14 +60,11 @@ io.on("connection", (socket) => {
 app.use(function (req, res, next) {
   res.header("Access-Control-Allow-Origin", "*");
   res.header("Access-Control-Allow-Methods", "GET, PUT, POST, DELETE");
-  res.header(
-    "Access-Control-Allow-Headers",
-    "Origin, X-Requested-With, Content-Type, Accept, Authorization"
-  );
+  res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, Authorization");
   next();
 });
 
-app.get("/clips", function (req, res) {
+app.get("/clips", API.authenticateKey, function (req, res) {
   let private = req.query.access == "private" ? 1 : 0;
   let query =
     "SELECT * from Clips WHERE private = " + private + " ORDER BY time DESC";
@@ -82,7 +87,14 @@ app.get("/clips", function (req, res) {
     });
 });
 
-app.delete("/clips", function (req, res) {
+app.delete("/clips", API.authenticateKey, function (req, res) {
+
+  if (!req.body.id) {
+    console.log("no id");
+    res.status(500).send('no id')
+    return;
+  }
+
   console.log(req.body);
   var query = "DELETE FROM Clips WHERE id = " + req.body.id + " ;";
   pool
@@ -100,7 +112,7 @@ app.delete("/clips", function (req, res) {
     });
 });
 
-app.post("/clips", function (req, res) {
+app.post("/clips", API.authenticateKey, function (req, res) {
   var query =
     "INSERT INTO Clips(content, time, private, imagesrc)" +
     "values('" +
@@ -150,7 +162,6 @@ app.post("/image", function (req, res) {
       console.log(err);
       return res.status(500).send({ msg: err + "Error occured" });
     }
-    // returing the response with file path and name
     console.log("file " + file.name + " downloaded");
     return res.send("file " + file.name + " uploaded");
   });
